@@ -1,24 +1,20 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using SixLabors.Fonts;
+using SixLabors.ImageSharp.Drawing.Processing;
+using SixLabors.ImageSharp.PixelFormats;
 using System.Diagnostics;
-using System.Drawing;
-using System.Drawing.Drawing2D;
-using System.Drawing.Imaging;
-using System.IO;
-using System.Text.Json;
 using System.Globalization;
-using System.Threading.Tasks;
-using System.Linq;
+using System.Text.Json;
+using Serilog;
 
 namespace libthumbnailer
 {
     public class ContactSheet
     {
-        public string FilePath { get; private set; }
+        public string FilePath { get; private set; } = string.Empty;
         public double Duration { get; private set; }
-        public string FileInfo { get; private set; }
-        public string AudioInfo { get; private set; }
-        public string VideoInfo { get; private set; }
+        public string FileInfo { get; private set; } = string.Empty;
+        public string AudioInfo { get; private set; } = string.Empty;
+        public string VideoInfo { get; private set; } = string.Empty;
         public int Rows { get; set; }
         public int Columns { get; set; }
         public int Width { get; set; }
@@ -26,35 +22,47 @@ namespace libthumbnailer
         public int Gap { get; set; }
         public List<Thumbnail> Thumbnails { get; }
 
-        private readonly Logger _logger;
         private readonly int _aspectRatio = 1;
-        private string _thumbDir;
+        private readonly Config _config;
+        private readonly ILogger _logger;
+        private string _thumbDir = string.Empty;
 
-        public static event EventHandler<string> SheetCreated;
-        public event EventHandler<string> SheetPrinted;
-        public static event EventHandler<string> AllSheetsPrinted;
+        public static event EventHandler<string>? SheetCreated;
+        public event EventHandler<string>? SheetPrinted;
+        //public static event EventHandler<string>? AllSheetsPrinted;
 
-        public ContactSheet(string filePath, Logger logger)
+        public ContactSheet(string path, Config config, ILogger logger)
         {
             CultureInfo.CurrentCulture = CultureInfo.GetCultureInfo("en-US");
+            FilePath = path;
+            _config = config;
             _logger = logger;
-            FilePath = filePath;
+
+            if (config is null || config.Rows == 0 || config.Columns == 0)
+            {
+                _logger.WithClassAndMethodNames<ContactSheet>().Warning("Config is invalid - loading default config");
+                config = Config.Load("default.json");
+            }
+
             var info = GetRootInfo();
             Duration = GetDuration(info);
             FileInfo = GetFileInfo(info);
             VideoInfo = GetVideoInfo(info);
             AudioInfo = GetAudioInfo(info);
 
-            Thumbnails = new List<Thumbnail>();
+            Rows = _config.Rows;
+            Columns = _config.Columns;
+            Width = _config.Width;
+            Gap = _config.Gap;
+
+            Thumbnails = [];
             Height = 0;
 
-            _logger.LogInfo($"--- Processing {filePath} ---");
+            _logger.WithClassAndMethodNames<ContactSheet>().Information("Initialized new contact sheet with W:{w}, H:{h}, G:{g}", Width, Height, Gap);
         }
 
         private JsonElement GetRootInfo()
         {
-            _logger.LogInfo($"Getting ffprobe info...");
-
             var probe = new Process
             {
                 StartInfo = new ProcessStartInfo
@@ -68,37 +76,34 @@ namespace libthumbnailer
             };
 
             string output;
-            try
-            {
-                probe.Start();
-                output = probe.StandardOutput.ReadToEnd().Trim();
-                probe.WaitForExit();
-                probe.Dispose();
-            }
-            catch
-            {
-                _logger.LogError($"ffprobe failed to start.");
-                throw new FfprobeException();
-            }
+
+            _logger.WithClassAndMethodNames<ContactSheet>().Information("Starting ffprobe");
+
+            var sw = new Stopwatch();
+
+            sw.Start();
+
+            probe.Start();
+            output = probe.StandardOutput.ReadToEnd().Trim();
+            probe.WaitForExit();
+            probe.Dispose();
+
+            sw.Stop();
+
+            _logger.WithClassAndMethodNames<ContactSheet>().Information("ffprobe finished in {time}ms", sw.ElapsedMilliseconds);
 
             JsonDocument doc = JsonDocument.Parse(output);
-            _logger.LogInfo($"Got ffprobe info");
             return doc.RootElement;
         }
 
         private double GetDuration(JsonElement root)
         {
-            try
-            {
-                return root.TryGetProperty("format", out var format) ? double.Parse(format.GetProperty("duration").GetString()) : -1;
-            }
-            catch
-            {
-                return -1;
-            }
+            var duration = root.TryGetProperty("format", out var format) ? double.Parse(format.GetProperty("duration").GetString()!) : -1;
+            _logger.WithClassAndMethodNames<ContactSheet>().Information("Got duration of {duration}", duration);
+            return duration;
         }
 
-        private string GetFileInfo(JsonElement root)
+        private static string GetFileInfo(JsonElement root)
         {
             try
             {
@@ -110,7 +115,7 @@ namespace libthumbnailer
             }
         }
 
-        private string GetVideoInfo(JsonElement root)
+        private static string GetVideoInfo(JsonElement root)
         {
             try
             {
@@ -123,7 +128,7 @@ namespace libthumbnailer
             }
         }
 
-        private string GetAudioInfo(JsonElement root)
+        private static string GetAudioInfo(JsonElement root)
         {
             try
             {
@@ -136,7 +141,7 @@ namespace libthumbnailer
             }
         }
 
-        private bool TryGetIndex(JsonElement streams, string streamName, out int index)
+        private static bool TryGetIndex(JsonElement streams, string streamName, out int index)
         {
             int len = streams.GetArrayLength();
             index = -1;
@@ -159,51 +164,68 @@ namespace libthumbnailer
         //Generates thumbnails by running ffmpeg on the file, and saving them in a temp folder.
         private void GenerateThumbnails()
         {
-            _logger.LogInfo($"Generating thumbnails...");
             double tween = Duration / (Rows * Columns);
 
             _thumbDir = "temp/temp_" + (FilePath.GetHashCode() + DateTime.Now.Millisecond);
             Directory.CreateDirectory(_thumbDir);
 
-            _logger.LogInfo($"Created temporary thumbnail folder {_thumbDir}");
-
-            _logger.LogInfo($"Running ffmpeg with arguments '-i \"{FilePath}\" -vf fps=1/{tween} {_thumbDir}/img%05d.png'...");
+            _logger.WithClassAndMethodNames<ContactSheet>().Information("tween is {tween} ({duration} / {rows}*{cols})", tween, Duration, Rows, Columns);
 
             var proc = new Process
             {
                 StartInfo = new ProcessStartInfo
                 {
                     FileName = "ffmpeg",
-                    Arguments = $"-i \"{FilePath}\" -vf fps=1/{tween} {_thumbDir}/img%05d.png",
+                    Arguments = $"-i \"{FilePath}\" -vf fps=1/{tween} {_thumbDir}/img%05d.jpeg",
                     UseShellExecute = false,
+                    RedirectStandardError = true,
                     CreateNoWindow = true
                 }
             };
 
+            _logger.WithClassAndMethodNames<ContactSheet>().Information("Starting ffmpeg with arguments {args}", proc.StartInfo.Arguments);
+
+            var sw = new Stopwatch();
+
+            sw.Start();
+
             proc.Start();
+            var output = proc.StandardError.ReadToEnd().Trim();
             proc.WaitForExit();
-            _logger.LogInfo($"ffmpeg exited with code {proc.ExitCode} at {proc.ExitTime}");
             proc.Dispose();
+
+            sw.Stop();
+
+            _logger.WithClassAndMethodNames<ContactSheet>().Information("ffmpeg finished in {time}ms", sw.ElapsedMilliseconds);
 
             int count = 0;
 
             var holder = Directory.GetFiles(_thumbDir);
+
+            if(holder.Length == 0)
+            {
+                _logger.WithClassAndMethodNames<ContactSheet>().Error("ffmpeg generated no images - Output:\n{out}", output);
+                throw new InvalidOperationException("No images were generated by ffmpeg");
+            }
+
             // Sort the list of screenshots lexicographically, because linux reads files weird, apparently...
             var sorted = from s in holder orderby s select s;
 
-            _logger.LogInfo($"Adding thumbnails...");
+            _logger.WithClassAndMethodNames<ContactSheet>().Information("Generated {n} pictures", holder.Length);
+
             foreach (string s in sorted)
             {
                 Thumbnails.Add(ThumbnailFactory.CreateThumbnail(s, ++count * tween));
+
+                _logger.WithClassAndMethodNames<ContactSheet>().Information("Creating thumbnail {n} of {total}", count, holder.Length);
             }
-            _logger.LogInfo($"Thumbnails done");
         }
 
         private string PrintInfo()
         {
-            return "File: " + new FileInfo(FilePath).Name + 
-                            Environment.NewLine + FileInfo + 
-                            Environment.NewLine + AudioInfo + 
+            return "File: " + new FileInfo(FilePath).Name +
+                            Environment.NewLine + FileInfo +
+                            Environment.NewLine + AudioInfo +
                             Environment.NewLine + VideoInfo;
         }
 
@@ -214,111 +236,76 @@ namespace libthumbnailer
         /// <param name="config">The configuration to use when printing.</param>
         /// <param name="overwrite">Whether or not to overwrite an existing file.</param>
         /// <returns>True if no errors occur; otherwise false.</returns>
-        public bool PrintSheet(string filename, Config config, bool overwrite)
+        public bool PrintSheet(bool overwrite)
         {
-            _logger.LogInfo($"Printing {filename}");
+            _logger.WithClassAndMethodNames<ContactSheet>().Information("Printing contact sheet");
 
-            if (!overwrite && File.Exists(filename + ".png"))
+            if (!overwrite && File.Exists(FilePath + ".png"))
             {
-                _logger.LogInfo($"SKIP: {filename} already exists - skipping");
                 return true;
             }
 
+            _logger.WithClassAndMethodNames<ContactSheet>().Information("Generating thumbnails");
             GenerateThumbnails();
 
-            _logger.LogInfo($"Generated {Thumbnails.Count} thumbnails");
-
-            _logger.LogInfo($"Defining dimensions and fonts...");
-            int imgWidth = (Width - ((Columns - 1) * Gap) - 4) / Columns;
+            int thumbWidth = (Width - ((Columns - 1) * Gap) - 4) / Columns;
             double tw = Thumbnails[0].Image.Width;
             double th = Thumbnails[0].Image.Height;
             double ratio = th / tw;
-            int imgHeight = (int)Math.Round(ratio * imgWidth / _aspectRatio);
-            Height = (imgHeight * Rows) + 2 + (Rows * Gap);
+            int thumbHeight = (int)Math.Round(ratio * thumbWidth / _aspectRatio);
+            Height = (thumbHeight * Rows) + 2 + (Rows * Gap);
             int infoHeight = 0;
 
-            Font infoF = FontFactory.CreateFont(Utils.GetFontFamilyFromName(config.InfoFont), config.InfoFontSize);
-            Font timeF = FontFactory.CreateFont(Utils.GetFontFamilyFromName(config.TimeFont), config.TimeFontSize);
-            SolidBrush infoB = BrushFactory.CreateBrush(Color.FromArgb(config.InfoFontColor));
-            SolidBrush timeB = BrushFactory.CreateBrush(Color.FromArgb(config.TimeFontColor));
-            SolidBrush timeSB = BrushFactory.CreateBrush(Color.FromArgb(config.ShadowColor));
+            Font infoF = new(Utils.GetFontFamilyFromName(_config.InfoFont), _config.InfoFontSize);
+            Font timeF = new(Utils.GetFontFamilyFromName(_config.TimeFont), _config.TimeFontSize);
+            SolidBrush infoB = BrushFactory.CreateBrush(Color.FromPixel(Rgba32.ParseHex(_config.InfoFontColor)));
+            SolidBrush timeB = BrushFactory.CreateBrush(Color.FromPixel(Rgba32.ParseHex(_config.TimeFontColor)));
+            SolidBrush timeSB = BrushFactory.CreateBrush(Color.FromPixel(Rgba32.ParseHex(_config.ShadowColor)));
 
-            if (config.PrintInfo)
+            if (_config.PrintInfo)
             {
-                infoHeight = Utils.GetStringHeight(PrintInfo(), infoF);
+                infoHeight = Utils.GetStringHeight(PrintInfo(), infoF) + 5;
                 Height += infoHeight;
             }
-            _logger.LogInfo($"Done!");
 
-            _logger.LogInfo($"Preparing bitmap...");
-            using (var bitmap = new Bitmap(Width, Height))
+            _logger.WithClassAndMethodNames<ContactSheet>().Information("Printing contact sheet");
+
+            using var image = new Image<Rgba32>(Width, Height);
+            image.Mutate(i => i.Fill(Color.Black));
+            int idx = 0;
+
+            for (int i = 0; i < Rows; i++)
             {
-                _logger.LogInfo($"Preparing canvas...");
-                using (var canvas = Graphics.FromImage(bitmap))
+                for (int j = 0; j < Columns; j++)
                 {
-                    canvas.InterpolationMode = InterpolationMode.HighQualityBicubic;
-                    canvas.SmoothingMode = SmoothingMode.HighQuality;
-                    canvas.TextRenderingHint = System.Drawing.Text.TextRenderingHint.AntiAlias;
-                    canvas.CompositingQuality = CompositingQuality.HighQuality;
-                    canvas.Clear(Color.FromArgb(config.BackgroundColor));
-                    int idx = 0;
+                    int x = 2 + (j * (thumbWidth + Gap));
+                    int y = infoHeight + (i * (thumbHeight + Gap));
 
-                    for (int i = 0; i < Rows; i++)
+                    var tn = Thumbnails[idx].Image;
+                    tn.Mutate(t => t.Resize(thumbWidth, thumbHeight));
+
+                    image.Mutate(i => i.DrawImage(tn, new Point(x, y), 1f));
+
+                    if (_config.PrintTime)
                     {
-                        for (int j = 0; j < Columns; j++)
-                        {
-                            int x = 2 + (j * (imgWidth + Gap));
-                            int y = infoHeight + (i * (imgHeight + Gap));
-                            try
-                            {
-                                canvas.DrawImage(Thumbnails[idx].Image, new Rectangle(x, y, imgWidth, imgHeight));
-                            }
-                            catch (Exception e)
-                            {
-                                _logger.LogWarning($"({FilePath}): {e.Message}");
-                                _logger.LogWarning($"Expecting {Rows * Columns} images, recieved {Thumbnails.Count}");
-                            }
-                            if (config.PrintTime)
-                            {
-                                string time = "";
-                                try
-                                {
-                                    time = Converter.ToHMS(Thumbnails[idx].TimeCode);
-                                }
-                                catch (Exception e)
-                                {
-                                    time = "?";
-                                    _logger.LogWarning($"TIMECODE ({FilePath}): {e.Message}");
-                                }
-                                int timeWidth = Utils.GetStringWidth(time, timeF, canvas);
-                                int timeHeight = Utils.GetStringHeight(time, timeF, canvas);
-                                canvas.DrawString(time, timeF, timeSB, new PointF(x + imgWidth - timeWidth + 1, y + imgHeight - timeHeight + 1));
-                                canvas.DrawString(time, timeF, timeB, new PointF(x + imgWidth - timeWidth, y + imgHeight - timeHeight));
-                            }
-                            idx++;
-                        }
+                        string time = Converter.ToHMS(Thumbnails[idx].TimeCode);
+
+                        int timeWidth = Utils.GetStringWidth(time, timeF);
+                        int timeHeight = Utils.GetStringHeight(time, timeF);
+                        image.Mutate(i => i.DrawText(time, timeF, timeSB, new PointF(x + thumbWidth - timeWidth + 1, y + thumbHeight - timeHeight + 1)));
+                        image.Mutate(i => i.DrawText(time, timeF, timeB, new PointF(x + thumbWidth - timeWidth, y + thumbHeight - timeHeight)));
                     }
-
-                    if (config.PrintInfo)
-                        canvas.DrawString(PrintInfo(), infoF, infoB, new PointF(2, 2));
-
-                    canvas.Save();
-                }
-                _logger.LogInfo($"Canvas done!");
-                try
-                {
-                    _logger.LogInfo($"Saving bitmap...");
-                    bitmap.Save(filename + ".png", ImageFormat.Png);
-                }
-                catch (Exception e)
-                {
-                    _logger.LogError($"Unable to save file {filename}: {e.Message}");
-                    return false;
+                    idx++;
                 }
             }
-            _logger.LogInfo($"Bitmap done!");
 
-            _logger.LogInfo($"Cleaning up...");
+            if (_config.PrintInfo)
+                image.Mutate(i => i.DrawText(PrintInfo(), infoF, infoB, new PointF(2, 2)));
+
+            image.SaveAsPng(FilePath + ".png");
+
+            _logger.WithClassAndMethodNames<ContactSheet>().Information("Cleaning up");
+
             foreach (var t in Thumbnails)
             {
                 t.Dispose();
@@ -328,7 +315,7 @@ namespace libthumbnailer
                 }
                 catch (Exception ex)
                 {
-                    _logger.LogWarning($"Failed to delete file {t.Path}: {ex.Message}");
+                    Console.WriteLine($"ERROR: {ex.Message}");
                 }
             }
             try
@@ -337,125 +324,25 @@ namespace libthumbnailer
             }
             catch (Exception ex)
             {
-                _logger.LogWarning($"Failed to delete directory {_thumbDir}: {ex.Message}");
+                Console.WriteLine($"ERROR: {ex.Message}");
             }
-            _logger.LogInfo($"Cleanup done!");
 
+            _logger.WithClassAndMethodNames<ContactSheet>().Information("Contact sheet printed successfully to {path}", FilePath + ".png");
             SheetPrinted?.Invoke(this, FilePath);
             return true;
         }
 
-        public static List<ContactSheet> BuildSheets(IEnumerable<string> files, Logger logger)
+        public List<ContactSheet> BuildSheets(IEnumerable<string> files)
         {
             var retval = new List<ContactSheet>();
 
             foreach (string file in files)
             {
-                try
-                {
-                    retval.Add(ContactSheetFactory.CreateContactSheet(file, logger));
-                    SheetCreated?.Invoke(null, file);
-                }
-                catch (Exception e)
-                {
-                    logger.LogError($"Exception during sheet-building of file {file}: {e.Message}");
-                    continue;
-                }
+                retval.Add(ContactSheetFactory.CreateContactSheet(file, _config, _logger));
+                SheetCreated?.Invoke(null, file);
             }
 
             return retval;
-        }
-
-        /// <summary>
-        /// Prints a list of contact sheets to file(s).
-        /// </summary>
-        /// <param name="sheets">The list of contact sheets to print.</param>
-        /// <param name="config">The configuration to use when printing.</param>
-        /// <param name="logger">The <see cref="Logger"/> instance to use for logging.</param>
-        /// <param name="overwrite">Whether or not to overwrite existing files.</param>
-        /// <param name="outputPath">The path to print files to, if it is different from the source path.</param>
-        /// <remarks>Prints the contact sheets sequentially.</remarks>
-        public static void PrintSheets(List<ContactSheet> sheets, Config config, Logger logger, bool overwrite, string outputPath = null)
-        {
-            var start = DateTime.Now;
-            var results = new List<bool>();
-
-            foreach (ContactSheet cs in sheets)
-            {
-                string filePath;
-                cs.Rows = Config.CurrentConfig.Rows;
-                cs.Columns = Config.CurrentConfig.Columns;
-                cs.Width = Config.CurrentConfig.Width;
-                cs.Gap = Config.CurrentConfig.Gap;
-
-                if (string.IsNullOrEmpty(outputPath))
-                {
-                    filePath = cs.FilePath;
-                }
-                else
-                {
-                    filePath = outputPath + "/" + new FileInfo(cs.FilePath).Name;
-                }
-
-                try
-                {
-                    results.Add(cs.PrintSheet(filePath, config, overwrite));
-                }
-                catch(Exception e)
-                {
-                    logger.LogError($"Exception caught while printing sheet: {filePath}: {e.Message}\n{e.StackTrace}");
-                    results.Add(false);
-                }
-            }
-
-            AllSheetsPrinted?.Invoke(null, "All done!");
-            logger.LogInfo($"*** Done in {DateTime.Now.Subtract(start).TotalSeconds} seconds ***");
-        }
-
-        /// <summary>
-        /// Prints a list of contact sheets to file(s).
-        /// </summary>
-        /// <param name="sheets">The list of contact sheets to print.</param>
-        /// <param name="config">The configuration to use when printing.</param>
-        /// <param name="logger">The <see cref="Logger"/> instance to use for logging.</param>
-        /// <param name="overwrite">Whether or not to overwrite existing files.</param>
-        /// <param name="outputPath">The path to print files to, if it is different from the source path.</param>
-        /// <remarks>Prints the contact sheets parallellized (multi-threaded).</remarks>
-        public static void PrintSheetsParallel(List<ContactSheet> sheets, Config config, Logger logger, bool overwrite, string outputPath = null)
-        {
-            var start = DateTime.Now;
-            List<bool> results = new List<bool>();
-
-            Parallel.ForEach(sheets, cs =>
-            {
-                string filePath;
-                cs.Rows = Config.CurrentConfig.Rows;
-                cs.Columns = Config.CurrentConfig.Columns;
-                cs.Width = Config.CurrentConfig.Width;
-                cs.Gap = Config.CurrentConfig.Gap;
-
-                if (string.IsNullOrEmpty(outputPath))
-                {
-                    filePath = cs.FilePath;
-                }
-                else
-                {
-                    filePath = outputPath + "/" + new FileInfo(cs.FilePath).Name;
-                }
-
-                try
-                {
-                    results.Add(cs.PrintSheet(filePath, config, overwrite));
-                }
-                catch (Exception e)
-                {
-                    logger.LogError($"Exceprion caught while printing sheet: {filePath}: {e.Message}");
-                    results.Add(false);
-                }
-            });
-
-            AllSheetsPrinted?.Invoke(null, "All done!");
-            logger.LogInfo($"*** Done in {DateTime.Now.Subtract(start).TotalSeconds} seconds ***");
         }
     }
 }
