@@ -1,10 +1,12 @@
-﻿namespace libthumbnailer
+﻿using Serilog;
+
+namespace libthumbnailer
 {
     public class ContactSheetFactory
     {
-        public static ContactSheet CreateContactSheet(string filePath, Logger logger)
+        public static ContactSheet CreateContactSheet(string filePath, Config config, ILogger logger)
         {
-            return new ContactSheet(filePath, logger);
+            return new ContactSheet(filePath, config, logger);
         }
     }
 }
