@@ -1,6 +1,4 @@
-﻿using System.Threading.Tasks;
-
-namespace libthumbnailer
+﻿namespace libthumbnailer
 {
     /// <summary>
     /// Factory methods for creating <see cref="Thumbnail"/> instances.
@@ -16,19 +14,6 @@ namespace libthumbnailer
         public static Thumbnail CreateThumbnail(string path, double timecode)
         {
             return new Thumbnail(path, timecode);
-        }
-
-        //not used - remove
-        public static Task<ContactSheet> CreateContactSheet(string path, int rows, int cols, int width, int gap, Logger logger)
-        {
-            return new Task<ContactSheet>(() =>
-                new ContactSheet(path, logger)
-                {
-                    Rows = rows,
-                    Columns = cols,
-                    Width = width,
-                    Gap = gap
-                });
         }
     }
 }
